fix: corrigindo tipagem incorreta de decimal para float

+++ schemas/production_line_schema.py
@@ -1,11 +1,10 @@
 from pydantic import BaseModel
-from decimal import Decimal
 
 
 class ProductionLineRequestSchema(BaseModel):
     name: str
     description: str
-    target_oee: Decimal
+    target_oee: float
     ideal_production_rate: float
 
 
@@ -13,7 +12,7 @@ class ProductionLineResponseSchema(BaseModel):
     id: int
     name: str
     description: str | None
-    target_oee: Decimal
+    target_oee: float
     ideal_production_rate: float
     current_production_rate: float
     production_rate: float
@@ -26,5 +25,5 @@ class ProductionLineResponseSchema(BaseModel):
 class ProductionLineUpdateSchema(BaseModel):
     name: str | None = None
     description: str | None = None
-    target_oee: Decimal | None = None
+    target_oee: float | None = None
     ideal_production_rate: float | None = None
